+++ server/index.js
@@ -8,8 +8,30 @@ const app = express();
 const db_url = process.env.MONGODB_URI
 const client_url = process.env.REACT_APP_CLIENT_URL
 const client_ws = process.env.REACT_APP_CLIENT_WS
-var status = 'fail'
+const WebSocket = require('ws');
+const http = require('http');
+const server = http.createServer(app);
+const wss = new WebSocket.Server({server})
+var Redis = require('ioredis');
+var redis = new Redis();
+var pub = new Redis();
 
+wss.on('connection', (ws) => {
+  ws.on('message',function (message){
+      ws.documentId = JSON.parse(message)
+      redis.subscribe('delete-list', function (err, count) {
+        if (err) console.error(err.message);
+    });
+  })
+  redis.on('message', function (channel, message) {
+    let listDetails = JSON.parse(message)
+    if(ws.documentId && ws.documentId.lists.length > 0 && ws.documentId.lists.filter((d)=>d.listName === listDetails.listName && d.details.ownedBy === listDetails.ownerEmail).length > 0){
+        const msg = Buffer.from(JSON.stringify({channel:channel,listName:listDetails.listName, ownedBy:listDetails.ownerEmail}));
+        ws.send(msg); 
+    }
+})
+});
+var status = 'fail'
 async function connect() {
   try {
     await mongoose.connect(db_url);
@@ -25,15 +47,14 @@ const corsOptions = {
 app.use(cors({options:corsOptions}))
 //app.options([client_ws], cors());
 connect()
-
-app.use('/api',cors({options:corsOptions}), routes);
+app.use('/api',cors({options:corsOptions}),routes);
 
 app.get('/', async(req,res) => {
   res.json({status:status})
 });
 
 const port = 8080;
-app.listen(port,(req,res) => {
+server.listen(port,(req,res) => {
   console.log(`Server listening on port ${port}`);
 });
